backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.1
groq>=0.4.2
python-dotenv>=1.0.0
pydantic>=2.5.3
sse-starlette>=1.8.2
